requirements.txt
fastapi
azure.cosmos
redis